requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.2.0
Flask-Cors==4.0.1
